+++ kdtree.h
@@ -4,7 +4,7 @@
 typedef void* Tree;
 typedef void* Info;
 
-Tree create_kdnode(int point[2], Info info);
-Tree insert_kd(Tree tree, int point[2], Info info, int depth);
-Tree insert_kd_init(Tree tree, int point[2], Info info);
+Tree create_kdnode(float point[2], Info info);
+Tree insert_kd(Tree tree, float point[2], Info info, int depth);
+Tree insert_kd_init(Tree tree, float point[2], Info info);
 void print(Tree tree);
